+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Changes
 
 - Added support for zipapps.
+- Added support for `PYTHONPATH`.
+- Added support for `-I` in Python 3.x variants.
 
 ## Version 12